Fixed ShibbolethLoginFilterIT test failures. The tests was copied to the ClarinShibbolethLoginFilter.
