+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/security/ShibbolethLoginFilterIT.java
@@ -75,58 +75,64 @@ public class ShibbolethLoginFilterIT extends AbstractControllerIntegrationTest {
         configurationService.setProperty("plugin.sequence.org.dspace.authenticate.AuthenticationMethod", SHIB_ONLY);
     }
 
-    @Test
-    public void testRedirectToDefaultDspaceUrl() throws Exception {
-        // NOTE: The initial call to /shibboleth comes *from* an external Shibboleth site. So, it is always
-        // unauthenticated, but it must include some expected SHIB attributes.
-        // SHIB-MAIL attribute is the default email header sent from Shibboleth after a successful login.
-        // In this test we are simply mocking that behavior by setting it to an existing EPerson.
-        String token = getClient().perform(get("/api/authn/shibboleth").requestAttr("SHIB-MAIL", eperson.getEmail()))
-                .andExpect(status().is3xxRedirection())
-                .andExpect(redirectedUrl("http://localhost:4000"))
-                .andReturn().getResponse().getHeader("Authorization");
+    // Note: This test was commented because the Shibboleth Authentication was customized following the Clarin
+    // requirements. This test was copied and updated following the Clarin updates to the
+    // `ClarinShibbolethLoginFilter#testRedirectToDefaultDspaceUrl` method.
+//    @Test
+//    public void testRedirectToDefaultDspaceUrl() throws Exception {
+//        // NOTE: The initial call to /shibboleth comes *from* an external Shibboleth site. So, it is always
+//        // unauthenticated, but it must include some expected SHIB attributes.
+//        // SHIB-MAIL attribute is the default email header sent from Shibboleth after a successful login.
+//        // In this test we are simply mocking that behavior by setting it to an existing EPerson.
+//        String token = getClient().perform(get("/api/authn/shibboleth").requestAttr("SHIB-MAIL", eperson.getEmail()))
+//                .andExpect(status().is3xxRedirection())
+//                .andExpect(redirectedUrl("http://localhost:4000"))
+//                .andReturn().getResponse().getHeader("Authorization");
+//
+//
+//        getClient(token).perform(get("/api/authn/status"))
+//                   .andExpect(status().isOk())
+//                   .andExpect(jsonPath("$.authenticated", is(true)))
+//                   .andExpect(jsonPath("$.authenticationMethod", is("shibboleth")));
+//
+//        getClient(token).perform(
+//                get("/api/authz/authorizations/search/object")
+//                        .param("embed", "feature")
+//                        .param("feature", feature)
+//                        .param("uri", utils.linkToSingleResource(ePersonRest, "self").getHref()))
+//                        .andExpect(status().isOk())
+//                        .andExpect(jsonPath("$.page.totalElements", is(0)))
+//                        .andExpect(jsonPath("$._embedded").doesNotExist());
+//
+//    }
 
-
-        getClient(token).perform(get("/api/authn/status"))
-                   .andExpect(status().isOk())
-                   .andExpect(jsonPath("$.authenticated", is(true)))
-                   .andExpect(jsonPath("$.authenticationMethod", is("shibboleth")));
-
-        getClient(token).perform(
-                get("/api/authz/authorizations/search/object")
-                        .param("embed", "feature")
-                        .param("feature", feature)
-                        .param("uri", utils.linkToSingleResource(ePersonRest, "self").getHref()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.page.totalElements", is(0)))
-                        .andExpect(jsonPath("$._embedded").doesNotExist());
-
-    }
-
-    @Test
-    public void testRedirectToGivenTrustedUrl() throws Exception {
-        String token = getClient().perform(get("/api/authn/shibboleth")
-                      .param("redirectUrl", "http://localhost:8080/server/api/authn/status")
-                      .requestAttr("SHIB-MAIL", eperson.getEmail()))
-                .andExpect(status().is3xxRedirection())
-                .andExpect(redirectedUrl("http://localhost:8080/server/api/authn/status"))
-                   .andReturn().getResponse().getHeader("Authorization");
-
-        getClient(token).perform(get("/api/authn/status"))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.authenticated", is(true)))
-                        .andExpect(jsonPath("$.authenticationMethod", is("shibboleth")));
-
-        getClient(token).perform(
-                get("/api/authz/authorizations/search/object")
-                        .param("embed", "feature")
-                        .param("feature", feature)
-                        .param("uri", utils.linkToSingleResource(ePersonRest, "self").getHref()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.page.totalElements", is(0)))
-                        .andExpect(jsonPath("$._embedded").doesNotExist());
-
-    }
+    // Note: This test was commented because the Shibboleth Authentication was customized following the Clarin
+    // requirements. This test was copied and updated following the Clarin updates to the
+    // `ClarinShibbolethLoginFilter#testRedirectToGivenTrustedUrl` method.
+//    @Test
+//    public void testRedirectToGivenTrustedUrl() throws Exception {
+//        String token = getClient().perform(get("/api/authn/shibboleth")
+//                      .param("redirectUrl", "http://localhost:8080/server/api/authn/status")
+//                      .requestAttr("SHIB-MAIL", eperson.getEmail()))
+//                .andExpect(status().is3xxRedirection())
+//                .andExpect(redirectedUrl("http://localhost:8080/server/api/authn/status"))
+//                   .andReturn().getResponse().getHeader("Authorization");
+//
+//        getClient(token).perform(get("/api/authn/status"))
+//                        .andExpect(status().isOk())
+//                        .andExpect(jsonPath("$.authenticated", is(true)))
+//                        .andExpect(jsonPath("$.authenticationMethod", is("shibboleth")));
+//
+//        getClient(token).perform(
+//                get("/api/authz/authorizations/search/object")
+//                        .param("embed", "feature")
+//                        .param("feature", feature)
+//                        .param("uri", utils.linkToSingleResource(ePersonRest, "self").getHref()))
+//                        .andExpect(status().isOk())
+//                        .andExpect(jsonPath("$.page.totalElements", is(0)))
+//                        .andExpect(jsonPath("$._embedded").doesNotExist());
+//
+//    }
 
     @Test
     public void testNoRedirectIfShibbolethDisabled() throws Exception {
@@ -141,27 +147,33 @@ public class ShibbolethLoginFilterIT extends AbstractControllerIntegrationTest {
                 .andExpect(status().isUnauthorized());
     }
 
-    @Test
-    public void testRedirectToAnotherGivenTrustedUrl() throws Exception {
-        String token = getAuthToken(eperson.getEmail(), password);
+    // Note: This test was commented because the Shibboleth Authentication was customized following the Clarin
+    // requirements. This test was copied and updated following the Clarin updates to the
+    // `ClarinShibbolethLoginFilter#testRedirectToAnotherGivenTrustedUrl` method.
+//    @Test
+//    public void testRedirectToAnotherGivenTrustedUrl() throws Exception {
+//        String token = getAuthToken(eperson.getEmail(), password);
+//
+//        getClient().perform(get("/api/authn/shibboleth")
+//                       .param("redirectUrl", "http://anotherdspacehost:4000/home")
+//                       .requestAttr("SHIB-MAIL", eperson.getEmail()))
+//                .andExpect(status().is3xxRedirection())
+//                .andExpect(redirectedUrl("http://anotherdspacehost:4000/home"));
+//    }
 
-        getClient().perform(get("/api/authn/shibboleth")
-                       .param("redirectUrl", "http://anotherdspacehost:4000/home")
-                       .requestAttr("SHIB-MAIL", eperson.getEmail()))
-                .andExpect(status().is3xxRedirection())
-                .andExpect(redirectedUrl("http://anotherdspacehost:4000/home"));
-    }
-
-    @Test
-    public void testRedirectToGivenUntrustedUrl() throws Exception {
-        // Now attempt to redirect to a URL that is NOT trusted (i.e. not in 'rest.cors.allowed-origins').
-
-        // Should result in a 400 error.
-        getClient().perform(get("/api/authn/shibboleth")
-                       .param("redirectUrl", "http://dspace.org")
-                       .requestAttr("SHIB-MAIL", eperson.getEmail()))
-                .andExpect(status().isBadRequest());
-    }
+    // Note: This test was commented because the Shibboleth Authentication was customized following the Clarin
+    // requirements. This test was copied and updated following the Clarin updates to the
+    // `ClarinShibbolethLoginFilter#testRedirectToGivenUntrustedUrl` method.
+//    @Test
+//    public void testRedirectToGivenUntrustedUrl() throws Exception {
+//        // Now attempt to redirect to a URL that is NOT trusted (i.e. not in 'rest.cors.allowed-origins').
+//
+//        // Should result in a 400 error.
+//        getClient().perform(get("/api/authn/shibboleth")
+//                       .param("redirectUrl", "http://dspace.org")
+//                       .requestAttr("SHIB-MAIL", eperson.getEmail()))
+//                .andExpect(status().isBadRequest());
+//    }
 
     @Test
     public void testNoRedirectIfInvalidShibAttributes() throws Exception {
@@ -187,44 +199,47 @@ public class ShibbolethLoginFilterIT extends AbstractControllerIntegrationTest {
                 .andExpect(status().isUnauthorized());
     }
 
-    @Test
-    public void patchPassword() throws Exception {
-
-        context.turnOffAuthorisationSystem();
-
-        EPerson ePerson = EPersonBuilder.createEPerson(context)
-                                        .withNameInMetadata("John", "Doe")
-                                        .withEmail("Johndoe@example.com")
-                                        .withPassword(password)
-                                        .build();
-
-        context.restoreAuthSystemState();
-
-        String newPassword = "newpassword";
-
-        List<Operation> ops = new ArrayList<Operation>();
-        AddOperation addOperation = new AddOperation("/password", newPassword);
-        ops.add(addOperation);
-        String patchBody = getPatchContent(ops);
-
-        // login through shibboleth
-        String token = getClient().perform(get("/api/authn/shibboleth").requestAttr("SHIB-MAIL", eperson.getEmail()))
-                                  .andExpect(status().is3xxRedirection())
-                                  .andExpect(redirectedUrl("http://localhost:4000"))
-                                  .andReturn().getResponse().getHeader("Authorization");
-
-
-        getClient(token).perform(get("/api/authn/status"))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.authenticated", is(true)))
-                        .andExpect(jsonPath("$.authenticationMethod", is("shibboleth")));
-
-        // updates password
-        getClient(token).perform(patch("/api/eperson/epersons/" + ePerson.getID())
-                                         .content(patchBody)
-                                         .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                        .andExpect(status().isForbidden());
-
-    }
+    // Note: This test was commented because the Shibboleth Authentication was customized following the Clarin
+    // requirements. This test was copied and updated following the Clarin updates to the
+    // `ClarinShibbolethLoginFilter#patchPassword` method.
+//    @Test
+//    public void patchPassword() throws Exception {
+//
+//        context.turnOffAuthorisationSystem();
+//
+//        EPerson ePerson = EPersonBuilder.createEPerson(context)
+//                                        .withNameInMetadata("John", "Doe")
+//                                        .withEmail("Johndoe@example.com")
+//                                        .withPassword(password)
+//                                        .build();
+//
+//        context.restoreAuthSystemState();
+//
+//        String newPassword = "newpassword";
+//
+//        List<Operation> ops = new ArrayList<Operation>();
+//        AddOperation addOperation = new AddOperation("/password", newPassword);
+//        ops.add(addOperation);
+//        String patchBody = getPatchContent(ops);
+//
+//        // login through shibboleth
+//        String token = getClient().perform(get("/api/authn/shibboleth").requestAttr("SHIB-MAIL", eperson.getEmail()))
+//                                  .andExpect(status().is3xxRedirection())
+//                                  .andExpect(redirectedUrl("http://localhost:4000"))
+//                                  .andReturn().getResponse().getHeader("Authorization");
+//
+//
+//        getClient(token).perform(get("/api/authn/status"))
+//                        .andExpect(status().isOk())
+//                        .andExpect(jsonPath("$.authenticated", is(true)))
+//                        .andExpect(jsonPath("$.authenticationMethod", is("shibboleth")));
+//
+//        // updates password
+//        getClient(token).perform(patch("/api/eperson/epersons/" + ePerson.getID())
+//                                         .content(patchBody)
+//                                         .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
+//                        .andExpect(status().isForbidden());
+//
+//    }
 
 }
